refactor: extract ToDisplayResult to converter function
Move ToDisplayResult method and related display conversion functions from
ExecutionContext to a separate converter package. This breaks the tight
coupling between execution types and display types.

Changes:
- Create pkg/ui/converter/execution_display.go with ConvertToDisplay function
- Move helper functions: mapOperationStatusToDisplayStatus, checkPackConfiguration, generateHandlerMessage
- Update all callers to use converter.ConvertToDisplay(ctx) instead of ctx.ToDisplayResult()
- Remove display-related imports from execution_context.go

This prepares for moving display types to pkg/ui/display in the next phase.

Part of #704
